microservices/aggregator-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.21
botocore==1.37.21
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
idna==3.10
jmespath==1.0.1
mysql-connector-python==9.2.0
numpy==2.2.4
pandas==2.2.3
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
sqlalchemy
httpx==0.28.1
requests
psycopg2-binary
mysql-connector-python 
python-dotenv